requirements.txt
agora-token-builder==1.0.0
asgiref==3.4.1
click==8.1.3
Django==4.0.1
django-crispy-forms==1.14.0
django-filter==23.1
django-queryset-csv==1.1.0
djangorestframework==3.14.0
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
imutils==0.5.4
numpy==1.24.1
opencv-python==4.7.0.72
Pillow==9.3.0
pytz==2023.3
scipy==1.10.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
unicodecsv==0.14.1
